delete movie in auth

+++ app/api/peliculas/[id]/route.tsx
@@ -1,14 +1,15 @@
 import { URL_API } from '@/config'
 import { NextRequest, NextResponse } from 'next/server'
-import datos from '../datos.json'
 
-/* export function GET(request: NextRequest, params: any) {
-  const pelicula = datos.peliculas.filter((peli) => peli.id === params.params.id)
-  if (!pelicula) return NextResponse.json({ message: 'Movie not finded' })
-  return NextResponse.json(pelicula, { status: 200 })
-} */
+export async function GET(request: NextRequest, params: any) {
+  return NextResponse.json(
+    { message: 'hello' },
+    {
+      status: 200,
+    }
+  )
+}
 export async function DELETE(request: NextRequest, params: any) {
-  const idMovie = params.params.id
   if (!request.cookies.get('token')) {
     return NextResponse.json(
       { message: `Not authorized` },
@@ -16,7 +17,8 @@ export async function DELETE(request: NextRequest, params: any) {
         status: 401,
       }
     )
-  } else {
+  }
+  try {
     const token = request.cookies.get('token') as { name: string; value: string }
     const rawData = await fetch(`${URL_API}/api/users/me`, {
       method: 'GET',
@@ -26,17 +28,28 @@ export async function DELETE(request: NextRequest, params: any) {
       },
     })
     const user = await rawData.json()
-    //TODO
-    const response = await fetch(`${URL_API}/api/asdf`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${token.value}`,
-      },
-    })
-    const data = await response.json()
-    return NextResponse.json(data, {
-      status: 200,
-    })
+    //DELETE MOVIE
+    if (user.id) {
+      const idMovie = params.params.id
+      const response = await fetch(`${URL_API}/api/peliculas/${idMovie}`, {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${token.value}`,
+        },
+      })
+      const data = await response.json()
+      return NextResponse.json(data, {
+        status: 200,
+      })
+    }
+    throw new Error('Ha ocurrido un error al intentar eliminar la pelicula')
+  } catch (error: any) {
+    return NextResponse.json(
+      { message: error.msg },
+      {
+        status: 401,
+      }
+    )
   }
 }
